requirements.txt
beautifulsoup4==4.11.1
black==22.3.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
google-api-core==2.7.3
google-auth==2.6.6
google-cloud==0.34.0
google-cloud-core==2.3.0
google-cloud-speech==2.13.1
googleapis-common-protos==1.56.0
grpcio==1.46.0
grpcio-status==1.46.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.20.0
requests==2.27.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
